move functions to paymentSession
